# Conflicts:
#	app/routes/guest_routes.py
#	app/templates/view_guest.html

+++ app/routes/guest_routes.py
@@ -12,8 +12,10 @@ from ..helpers import (
     get_form_value,
     generate_guest_number, user_has_access, is_different
 )
-from ..models import db as sqlalchemy_db, Guest, Animal, Payments, Representative, ChangeLog, FoodHistory, FieldRegistry
+from ..models import db as sqlalchemy_db, Guest, Animal, Payments, Representative, ChangeLog, FoodHistory, FoodTag, \
+    FieldRegistry
 from ..reports import generate_gast_card_pdf
+from sqlalchemy.sql.expression import func
 
 guest_bp = Blueprint("guest", __name__)
 
@@ -63,14 +65,16 @@ def view_guest(guest_id):
         changelog = (
             ChangeLog.query.filter_by(guest_id=guest.id)
             .order_by(ChangeLog.change_timestamp.desc())
-            .limit(10)
             .all()
         )
         payments = (
             Payments.query.filter_by(guest_id=guest.id)
             .order_by(Payments.created_on.desc())
-            .limit(10)
             .all()
+        )
+
+        all_tags = (
+            FoodTag.query.all()
         )
         # Build ordered list of visible Guest fields with UI labels
         all_fields = FieldRegistry.query.filter_by(model_name="Guest").all()
@@ -109,6 +113,7 @@ def view_guest(guest_id):
     else:
         animals = []
         changelog = []
+        all_tags = []
         feed_history = []
         payments = []
         visible_fields_guest = []
@@ -120,6 +125,7 @@ def view_guest(guest_id):
             visible_fields_guest=visible_fields_guest,
             visible_fields_animal=visible_fields_animal,
             guest=guest,
+            all_tags=all_tags,
             representative=representative,
             animals=animals,
             changelog=changelog,
@@ -435,11 +441,13 @@ def activate_guest(guest_id):
 @roles_required("admin")
 @login_required
 def delete_guest(guest_id):
-    Guest.query.filter_by(id=guest_id).delete()
+    if Payments.query.filter_by(guest_id=guest_id).first():
+        flash("Gast ist Buchalterisch relevant und kann nicht gelöscht werden.", "danger")
+        return redirect(url_for("guest.list_guests"))
     Animal.query.filter_by(guest_id=guest_id).delete()
     FoodHistory.query.filter_by(guest_id=guest_id).delete()
     ChangeLog.query.filter_by(guest_id=guest_id).delete()
-    Payments.query.filter_by(guest_id=guest_id).delete()
+    Guest.query.filter_by(id=guest_id).delete()
     sqlalchemy_db.session.commit()
     session["guests_changed"] = True
     flash("Gast wurde vollständig gelöscht.", "success")
